Removed circular dependencies between files (flatten dependency graph), added Madge as part of precommit hook to avoid regression.
