+++ packages/inferno-router/src/Router.ts
@@ -7,17 +7,7 @@ import Component from "inferno-component";
 import VNodeFlags from "inferno-vnode-flags";
 import match, { matchPath } from "./match";
 import RouterContext from "./RouterContext";
-
-export interface IRouterProps {
-  history?: any;
-  children?: any;
-  router: any;
-  location: any;
-  baseUrl?: any;
-  component?: Component<any, any>;
-  asyncBefore?: any;
-  onUpdate?: any;
-}
+import { IRouterProps } from "./routerprops";
 
 function createrRouter(history) {
   if (!history) {
