+++ packages/inferno/src/DOM/patching.ts
@@ -5,36 +5,24 @@
 import {
   combineFrom,
   isArray,
-  isFunction,
   isInvalid,
   isNull,
   isNullOrUndef,
-  isNumber,
   isObject,
-  isString,
   isStringOrNumber,
   isUndefined,
   NO_OP,
   throwError
 } from "inferno-shared";
 import VNodeFlags from "inferno-vnode-flags";
-import { options } from "../core/options";
 import {
   createTextVNode,
   createVoidVNode,
   directClone,
   isVNode,
+  options,
   VNode
-} from "../core/VNodes";
-import {
-  booleanProps,
-  delegatedEvents,
-  isUnitlessNumber,
-  namespaces,
-  skipProps,
-  strictProps
-} from "./constants";
-import { handleEvent } from "./events/delegation";
+} from "../core/implementation";
 import {
   mount,
   mountArrayChildren,
@@ -44,25 +32,54 @@ import {
   mountText,
   mountVoid
 } from "./mounting";
-import { componentToDOMNodeMap } from "./rendering";
 import { unmount } from "./unmounting";
 import {
   appendChild,
+  componentToDOMNodeMap,
   EMPTY_OBJ,
   insertOrAppend,
   isKeyed,
-  isSameInnerHTML,
-  removeAllChildren,
   replaceChild,
-  replaceVNode,
-  replaceWithNewNode,
   setTextContent,
   updateTextContent
-} from "./utils";
+} from "./utils/common";
 import {
   isControlledFormElement,
   processElement
 } from "./wrappers/processElement";
+import { patchProp, removeProp } from "./props";
+
+function replaceVNode(parentDom, dom, vNode) {
+  unmount(vNode, null);
+  replaceChild(parentDom, dom, vNode.dom);
+}
+
+function removeAllChildren(dom: Element, children) {
+  for (let i = 0, len = children.length; i < len; i++) {
+    const child = children[i];
+
+    if (!isInvalid(child)) {
+      unmount(child, null);
+    }
+  }
+  dom.textContent = "";
+}
+
+function replaceWithNewNode(
+  lastNode,
+  nextNode,
+  parentDom,
+  lifecycle: Function[],
+  context: Object,
+  isSVG: boolean
+) {
+  unmount(lastNode, null);
+  replaceChild(
+    parentDom,
+    mount(nextNode, null, lifecycle, context, isSVG),
+    lastNode.dom
+  );
+}
 
 export function patch(
   lastVNode: VNode,
@@ -117,21 +134,13 @@ export function patch(
       if (lastFlags & VNodeFlags.Text) {
         patchText(lastVNode, nextVNode);
       } else {
-        replaceVNode(
-          parentDom,
-          mountText(nextVNode, null),
-          lastVNode
-        );
+        replaceVNode(parentDom, mountText(nextVNode, null), lastVNode);
       }
     } else if (nextFlags & VNodeFlags.Void) {
       if (lastFlags & VNodeFlags.Void) {
         patchVoid(lastVNode, nextVNode);
       } else {
-        replaceVNode(
-          parentDom,
-          mountVoid(nextVNode, null),
-          lastVNode
-        );
+        replaceVNode(parentDom, mountVoid(nextVNode, null), lastVNode);
       }
     } else {
       // Error case: mount new one replacing old one
@@ -144,10 +153,7 @@ export function patch(
   }
 }
 
-function unmountChildren(
-  children,
-  dom: Element
-) {
+function unmountChildren(children, dom: Element) {
   if (isVNode(children)) {
     unmount(children, dom);
   } else if (isArray(children)) {
@@ -324,14 +330,7 @@ function patchChildren(
     mount(nextChildren, dom, lifecycle, context, isSVG);
   } else if (isVNode(nextChildren)) {
     if (isVNode(lastChildren)) {
-      patch(
-        lastChildren,
-        nextChildren,
-        dom,
-        lifecycle,
-        context,
-        isSVG
-      );
+      patch(lastChildren, nextChildren, dom, lifecycle, context, isSVG);
     } else {
       unmountChildren(lastChildren, dom);
       mount(nextChildren, dom, lifecycle, context, isSVG);
@@ -555,14 +554,7 @@ export function patchComponent(
           }
         }
         if (nextInput !== NO_OP) {
-          patch(
-            lastInput,
-            nextInput,
-            parentDom,
-            lifecycle,
-            context,
-            isSVG
-          );
+          patch(lastInput, nextInput, parentDom, lifecycle, context, isSVG);
           nextVNode.children = nextInput;
           if (
             nextHooksDefined &&
@@ -619,14 +611,7 @@ export function patchNonKeyedChildren(
     if (nextChild.dom) {
       nextChild = nextChildren[i] = directClone(nextChild);
     }
-    patch(
-      lastChildren[i],
-      nextChild,
-      dom,
-      lifecycle,
-      context,
-      isSVG
-    );
+    patch(lastChildren[i], nextChild, dom, lifecycle, context, isSVG);
   }
   if (lastChildrenLength < nextChildrenLength) {
     for (i = commonLength; i < nextChildrenLength; i++) {
@@ -681,14 +666,7 @@ export function patchKeyedChildren(
   outer: {
     // Sync nodes with the same key at the beginning.
     while (aStartNode.key === bStartNode.key) {
-      patch(
-        aStartNode,
-        bStartNode,
-        dom,
-        lifecycle,
-        context,
-        isSVG
-      );
+      patch(aStartNode, bStartNode, dom, lifecycle, context, isSVG);
       aStart++;
       bStart++;
       if (aStart > aEnd || bStart > bEnd) {
@@ -936,140 +914,4 @@ function lis_algorithm(arr: number[]): number[] {
   }
 
   return result;
-}
-
-export function isAttrAnEvent(attr: string): boolean {
-  return attr[0] === "o" && attr[1] === "n";
-}
-
-export function patchProp(
-  prop,
-  lastValue,
-  nextValue,
-  dom: Element,
-  isSVG: boolean,
-  hasControlledValue: boolean
-) {
-  if (lastValue !== nextValue) {
-    if (skipProps.has(prop) || (hasControlledValue && prop === "value")) {
-      return;
-    } else if (booleanProps.has(prop)) {
-      prop = prop === "autoFocus" ? prop.toLowerCase() : prop;
-      dom[prop] = !!nextValue;
-    } else if (strictProps.has(prop)) {
-      const value = isNullOrUndef(nextValue) ? "" : nextValue;
-
-      if (dom[prop] !== value) {
-        dom[prop] = value;
-      }
-    } else if (isAttrAnEvent(prop)) {
-      patchEvent(prop, lastValue, nextValue, dom);
-    } else if (isNullOrUndef(nextValue)) {
-      dom.removeAttribute(prop);
-    } else if (prop === "style") {
-      patchStyle(lastValue, nextValue, dom);
-    } else if (prop === "dangerouslySetInnerHTML") {
-      const lastHtml = lastValue && lastValue.__html;
-      const nextHtml = nextValue && nextValue.__html;
-
-      if (lastHtml !== nextHtml) {
-        if (!isNullOrUndef(nextHtml) && !isSameInnerHTML(dom, nextHtml)) {
-          dom.innerHTML = nextHtml;
-        }
-      }
-    } else {
-      // We optimize for NS being boolean. Its 99.9% time false
-      if (isSVG && namespaces.has(prop)) {
-        // If we end up in this path we can read property again
-        dom.setAttributeNS(namespaces.get(prop) as string, prop, nextValue);
-      } else {
-        dom.setAttribute(prop, nextValue);
-      }
-    }
-  }
-}
-
-export function patchEvent(name: string, lastValue, nextValue, dom) {
-  if (lastValue !== nextValue) {
-    if (delegatedEvents.has(name)) {
-      handleEvent(name, lastValue, nextValue, dom);
-    } else {
-      const nameLowerCase = name.toLowerCase();
-      const domEvent = dom[nameLowerCase];
-      // if the function is wrapped, that means it's been controlled by a wrapper
-      if (domEvent && domEvent.wrapped) {
-        return;
-      }
-      if (!isFunction(nextValue) && !isNullOrUndef(nextValue)) {
-        const linkEvent = nextValue.event;
-
-        if (linkEvent && isFunction(linkEvent)) {
-          dom[nameLowerCase] = function(e) {
-            linkEvent(nextValue.data, e);
-          };
-        } else {
-          if (process.env.NODE_ENV !== "production") {
-            throwError(
-              `an event on a VNode "${name}". was not a function or a valid linkEvent.`
-            );
-          }
-          throwError();
-        }
-      } else {
-        dom[nameLowerCase] = nextValue;
-      }
-    }
-  }
-}
-
-// We are assuming here that we come from patchProp routine
-// -nextAttrValue cannot be null or undefined
-function patchStyle(lastAttrValue, nextAttrValue, dom) {
-  const domStyle = dom.style;
-  let style;
-  let value;
-
-  if (isString(nextAttrValue)) {
-    domStyle.cssText = nextAttrValue;
-    return;
-  }
-
-  if (!isNullOrUndef(lastAttrValue) && !isString(lastAttrValue)) {
-    for (style in nextAttrValue) {
-      // do not add a hasOwnProperty check here, it affects performance
-      value = nextAttrValue[style];
-      if (value !== lastAttrValue[style]) {
-        domStyle[style] =
-          !isNumber(value) || isUnitlessNumber.has(style)
-            ? value
-            : value + "px";
-      }
-    }
-
-    for (style in lastAttrValue) {
-      if (isNullOrUndef(nextAttrValue[style])) {
-        domStyle[style] = "";
-      }
-    }
-  } else {
-    for (style in nextAttrValue) {
-      value = nextAttrValue[style];
-      domStyle[style] =
-        !isNumber(value) || isUnitlessNumber.has(style) ? value : value + "px";
-    }
-  }
-}
-
-function removeProp(prop: string, lastValue, dom, nextFlags: number) {
-  if (prop === "value") {
-    // When removing value of select element, it needs to be set to null instead empty string, because empty string is valid value for option which makes that option selected
-    // MS IE/Edge don't follow html spec for textArea and input elements and we need to set empty string to value in those cases to avoid "null" and "undefined" texts
-    dom.value = nextFlags & VNodeFlags.SelectElement ? null : "";
-  } else if (prop === "style") {
-    dom.removeAttribute("style");
-  } else if (isAttrAnEvent(prop)) {
-    handleEvent(prop, lastValue, null, dom);
-  } else {
-    dom.removeAttribute(prop);
-  }
 }
