+++ packages/inferno/src/DOM/unmounting.ts
@@ -8,20 +8,15 @@ import {
   isInvalid,
   isNull,
   isNullOrUndef,
-  isObject,
+  isObject
 } from "inferno-shared";
 import VNodeFlags from "inferno-vnode-flags";
-import { options } from "../core/options";
-import { VNode } from "../core/VNodes";
-import {delegatedEvents} from "./constants";
-import {handleEvent} from "./events/delegation";
-import { componentToDOMNodeMap } from "./rendering";
-import { EMPTY_OBJ, removeChild } from "./utils";
+import { VNode, options } from "../core/implementation";
+import { delegatedEvents } from "./constants";
+import { handleEvent } from "./events/delegation";
+import { EMPTY_OBJ, removeChild, componentToDOMNodeMap } from "./utils/common";
 
-export function unmount(
-  vNode: VNode,
-  parentDom: Element | null
-) {
+export function unmount(vNode: VNode, parentDom: Element | null) {
   const flags = vNode.flags;
   const dom = vNode.dom as Element;
 
@@ -32,36 +27,33 @@ export function unmount(
     const props = vNode.props || EMPTY_OBJ;
     const ref = vNode.ref as any;
 
-      if (isStatefulComponent) {
-        if (!instance._unmounted) {
-          if (isFunction(options.beforeUnmount)) {
-            options.beforeUnmount(vNode);
-          }
-          if (isFunction(instance.componentWillUnmount)) {
-            instance.componentWillUnmount();
-          }
-          if (isFunction(ref)) {
-            ref(null);
-          }
-          instance._unmounted = true;
-          if (options.findDOMNodeEnabled) {
-            componentToDOMNodeMap.delete(instance);
-          }
-
-          unmount(
-            instance._lastInput,
-            null
-          );
+    if (isStatefulComponent) {
+      if (!instance._unmounted) {
+        if (isFunction(options.beforeUnmount)) {
+          options.beforeUnmount(vNode);
         }
-      } else {
-        if (!isNullOrUndef(ref)) {
-          if (isFunction(ref.onComponentWillUnmount)) {
-            ref.onComponentWillUnmount(dom, props);
-          }
+        if (isFunction(instance.componentWillUnmount)) {
+          instance.componentWillUnmount();
+        }
+        if (isFunction(ref)) {
+          ref(null);
+        }
+        instance._unmounted = true;
+        if (options.findDOMNodeEnabled) {
+          componentToDOMNodeMap.delete(instance);
         }
 
-        unmount(instance, null);
+        unmount(instance._lastInput, null);
       }
+    } else {
+      if (!isNullOrUndef(ref)) {
+        if (isFunction(ref.onComponentWillUnmount)) {
+          ref.onComponentWillUnmount(dom, props);
+        }
+      }
+
+      unmount(instance, null);
+    }
   } else if ((flags & VNodeFlags.Element) > 0) {
     const ref = vNode.ref as any;
     const props = vNode.props;
@@ -94,7 +86,7 @@ export function unmount(
       for (const name in props) {
         // Remove all delegated events, regular events die with dom node
         if (delegatedEvents.has(name)) {
-          handleEvent(name, props[name], null, dom)
+          handleEvent(name, props[name], null, dom);
         }
       }
     }
