+++ packages/inferno/src/DOM/utils/innerhtml.ts
@@ -1,0 +1,19 @@
+/**
+ * @module Inferno
+ */ /** TypeDoc Comment */
+
+export function isSameInnerHTML(dom: Element, innerHTML: string): boolean {
+  const tempdom = document.createElement("i");
+
+  tempdom.innerHTML = innerHTML;
+  return tempdom.innerHTML === dom.innerHTML;
+}
+
+export function isSamePropsInnerHTML(dom: Element, props): boolean {
+  return Boolean(
+    props &&
+      props.dangerouslySetInnerHTML &&
+      props.dangerouslySetInnerHTML.__html &&
+      isSameInnerHTML(dom, props.dangerouslySetInnerHTML.__html)
+  );
+}
