+++ packages/inferno/src/DOM/wrappers/SelectWrapper.ts
@@ -3,9 +3,9 @@
  */ /** TypeDoc Comment */
 
 import { isArray, isInvalid, isNullOrUndef } from "inferno-shared";
-import { isVNode } from "../../core/VNodes";
-import { EMPTY_OBJ } from "../utils";
-import {createWrappedFunction} from "./wrapper";
+import { isVNode } from "../../core/implementation";
+import { EMPTY_OBJ } from "../utils/common";
+import { createWrappedFunction } from "./wrapper";
 
 function updateChildOptionGroup(vNode, value) {
   const type = vNode.type;
@@ -41,7 +41,7 @@ function updateChildOption(vNode, value) {
   }
 }
 
-const onSelectChange = createWrappedFunction('onChange', applyValue);
+const onSelectChange = createWrappedFunction("onChange", applyValue);
 
 export function processSelect(
   vNode,
