--- packages/inferno/src/core/options.ts
@@ -1,30 +1,0 @@
-/**
- * @module Inferno
- */ /** TypeDoc Comment */
-
-import {VNode} from "./VNodes";
-
-export interface Root {
-  dom: Element | SVGAElement;
-  input: VNode;
-}
-
-export const options: {
-  afterMount: null | Function;
-  afterRender: null | Function;
-  afterUpdate: null | Function;
-  beforeRender: null | Function;
-  beforeUnmount: null | Function;
-  createVNode: null | Function;
-  findDOMNodeEnabled: boolean;
-  roots: Root[];
-} = {
-  afterMount: null,
-  afterRender: null,
-  afterUpdate: null,
-  beforeRender: null,
-  beforeUnmount: null,
-  createVNode: null,
-  findDOMNodeEnabled: false,
-  roots: []
-};
